Refactored to break out components in elementCreation to separate modules.
